support for default domain in jumpgate.conf
adds support for the specification of the default domain to use when
booting servers in the conf file rather than hardcoding it to
'jumpgate.com' in the py code. this is a partial implementation of
https://github.com/softlayer/jumpgate/issues/63 which allows
configurable domain per jumpgate service. a future PR will include
additional support such as that discussed in
https://github.com/softlayer/jumpgate/issues/63