Adding token validation api for v3 identity api
This change impleneted token valication api which is a major function as
a part of v3 identity api, some upper layer components used this
interface to validate v3 token.

An action handling function has been added which is used to handle the
http request received on /v3/auth/tokens endpoint with GET method.

Signed-off-by: Zhi Yan Liu <zhiyanl@cn.ibm.com>